feat : #54 - 싱글게임 API 정합성 검사

+++ backend/src/main/java/com/backend/api/domain/single/dto/response/StockChartDto.java
@@ -14,7 +14,7 @@ public record StockChartDto(
     @Schema(description = "종가")
     Integer endPrice,
     @Schema(description = "거래량")
-    Integer tradingVolume,
+    Long tradingVolume,
 
     @Schema(description = "날짜")
     LocalDateTime date
